Swift 3,2 compatibility. Cast value to UInt, as Xcode9b5 compiler appears to default value to an Int and then errors on mixing UInt and Int operations. Accept all Xcode9 suggested compiler warning settings. Set SWIFT_VERSION to 3.2 in all targets. All targets still build with Swift 3.1/Xcode 8.3.x. Rename Network example framework to WebServices to avoid a name collision with an iOS 11 private framework with the same name.
