+++ Example/Frameworks/WebServices.xcodeproj/project.xcworkspace/contents.xcworkspacedata
